Use int64_t for LaneInUse
Downstream uses int64_t for LanesInUse,
but upstream uses size_t.

Related links are:

Downstream:
----------

https://github.ibm.com/openbmc/phosphor-dbus-interfaces/blob/dd8ca6c72c1fc4eee5b4bfc2904474b796c5bd01/yaml/xyz/openbmc_project/Inventory/Item/PCIeDevice.interface.yaml#L291

https://github.ibm.com/openbmc/phosphor-dbus-interfaces/blob/dd8ca6c72c1fc4eee5b4bfc2904474b796c5bd01/yaml/xyz/openbmc_project/Inventory/Item/PCIeDevice.interface.yaml#L291

https://github.ibm.com/openbmc/pldm/blob/48ecf0481e3a42dfc59e7dafff91f3197fba1fa2/host-bmc/dbus/pcie_device.hpp#L42

ibm-openbmc:
-----------

https://github.com/ibm-openbmc/pldm/blob/3712d289ba04b9c7a994c203a0a674d1270c45f1/host-bmc/dbus/pcie_device.hpp#L42

Upstream:
---------
https://github.com/search?q=repo%3Aopenbmc%2Fphosphor-dbus-interfaces%20LanesInUse&type=code

https://github.com/openbmc/phosphor-dbus-interfaces/blob/4b6755c9894cf5d4e983c6e3b86fc752695a1d8e/yaml/xyz/openbmc_project/Inventory/Item/PCIeDevice.interface.yaml#L292

```
[ERROR dbus_utils.hpp:17] DBUS property error in property: LanesInUse, reason: 1  <-- WrongType
[CRITICAL error_messages.cpp:284] Internal Error /usr/src/debug/bmcweb/1.0+git-r0/redfish-core/lib/pcie.hpp(431:32) `void redfish::addPCIeDeviceProperties(const std::shared_ptr<bmcweb::AsyncResp>&, const std::string&, const dbus::utility::DBusPropertiesMap&)`:
```

```
.LanesInUse                                           property  x         0
```

```
[DEBUG routing.hpp:669] Matched rule '/redfish/v1/Systems/<str>/PCIeDevices/<str>/' 1 / 2
[DEBUG dbus_privileges.hpp:53] userName = admin userRole = priv-admin
[DEBUG query.hpp:121] setup redfish route
[DEBUG http_response.hpp:309] 0x158b418 releasing completion handlertrue
[DEBUG http_response.hpp:299] 0x158b418 setting completion handler
[ERROR dbus_utils.hpp:17] DBUS property error in property: LanesInUse, reason: 1
[CRITICAL error_messages.cpp:284] Internal Error /usr/src/debug/bmcweb/1.0+git-r0/redfish-core/lib/pcie.hpp(431:32) `void redfish::addPCIeDeviceProperties(const std::shared_ptr<bmcweb::AsyncResp>&, const std::string&, const dbus::utility::DBusPropertiesMap&)`:
[DEBUG pcie.hpp:253] PCIeDevice is not associated with PCIeSlot
[DEBUG http_response.hpp:284] 0x158b418 calling completion handler
[DEBUG http_response.hpp:287] 0x158b418 completion handler was valid
[DEBUG query_param.hpp:1019] Processing query params
[DEBUG http_response.hpp:86] Moving response containers; this: 0x16c70d0; other: 0x158b418
[INFO complete_response_fields.hpp:23] Response:  /redfish/v1/Systems/system/PCIeDevices/pcie_card10 500
[DEBUG http_connection.hpp:593] 0x16c4ee0 doWrite
[DEBUG http_connection.hpp:651] 0x16c4ee0 timer started
[DEBUG http_response.hpp:299] 0x16c70d0 setting completion handler
```

+++ redfish-core/lib/pcie.hpp
@@ -418,7 +418,7 @@ inline void addPCIeDeviceProperties(
 {
     const std::string* generationInUse = nullptr;
     const std::string* generationSupported = nullptr;
-    const size_t* lanesInUse = nullptr;
+    const int64_t * lanesInUse = nullptr;
     const size_t* maxLanes = nullptr;
 
     const bool success = sdbusplus::unpackPropertiesNoThrow(
